#10242: Migrate tanh_bw to TTNN
