+++ docs/source/ttnn/ttnn/api.rst
@@ -179,6 +179,7 @@ Pointwise Unary
    ttnn/bias_gelu_bw
    ttnn/pow_bw
    ttnn/exp_bw
+   ttnn/tanh_bw
    ttnn/assign_bw
    ttnn/multigammaln_bw
    ttnn/add_bw
