+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -653,8 +653,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.sqrt_bw
 
-.. autofunction:: tt_lib.tensor.tanh_bw
-
 .. autofunction:: tt_lib.tensor.complex_abs_bw
 
 .. autofunction:: tt_lib.tensor.gelu_bw
