+++ docs/source/ttnn/ttnn/ttnn/tanh_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.tanh_bw:
+
+ttnn.tanh_bw
+###############
+
+.. autofunction:: ttnn.tanh_bw
