+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -36,22 +36,6 @@ std::vector<std::optional<Tensor>> sqrt_bw(
     const std::vector<bool>& are_required_outputs = std::vector<bool>{true},
     std::optional<Tensor> input_grad = std::nullopt);
 
-// bw = grad(1 - tanh(x) ** 2)
-std::vector<std::optional<Tensor>> tanh_bw(
-    uint8_t cq_id,
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-    const std::vector<bool>& are_required_outputs = std::vector<bool>{true},
-    std::optional<Tensor> input_grad = std::nullopt);
-
-std::vector<std::optional<Tensor>> tanh_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG,
-    const std::vector<bool>& are_required_outputs = std::vector<bool>{true},
-    std::optional<Tensor> input_grad = std::nullopt);
-
 std::vector<Tensor> binary_le_bw(
     const Tensor& grad,
     const Tensor& input,
