+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -23,6 +23,7 @@ enum class UnaryBackwardOpType {
     RDIV_BW,
     BIAS_GELU_BW,
     POW_BW,
+    TANH_BW,
     EXP_BW,
     ASSIGN_BW,
     MULTIGAMMALN_BW,
@@ -121,6 +122,7 @@ std::vector<std::optional<Tensor>> _pow_bw(uint8_t queue_id, const Tensor& grad,
 
 //OpHandler_unary_optional : get_function_unary_optional
 std::vector<std::optional<Tensor>> _exp_bw(uint8_t queue_id, const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config, const std::vector<bool>& are_required_outputs, std::optional<Tensor> input_grad);
+std::vector<std::optional<Tensor>> _tanh_bw(uint8_t queue_id, const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config, const std::vector<bool>& are_required_outputs, std::optional<Tensor> input_grad);
 
 // OpHandler struct template
 template <UnaryBackwardOpType OpType>
@@ -198,6 +200,13 @@ template <>
 struct OpHandler_unary_optional<UnaryBackwardOpType::EXP_BW> {
     static std::vector<std::optional<Tensor>> handle( uint8_t queue_id, const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config, const std::vector<bool>& are_required_outputs, std::optional<Tensor> input_grad ) {
         return _exp_bw(queue_id, grad, input, output_mem_config, are_required_outputs, input_grad);
+    }
+};
+
+template <>
+struct OpHandler_unary_optional<UnaryBackwardOpType::TANH_BW> {
+    static std::vector<std::optional<Tensor>> handle( uint8_t queue_id, const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config, const std::vector<bool>& are_required_outputs, std::optional<Tensor> input_grad ) {
+        return _tanh_bw(queue_id, grad, input, output_mem_config, are_required_outputs, input_grad);
     }
 };
